requirements.txt
tensorflow-cpu
pymongo
flask
